+++ packages/api/src/hooks/index.ts
@@ -5,6 +5,7 @@ export { default as useAccountsList } from './useAccountsList';
 export { default as useActiveAccount } from './useActiveAccount';
 export { default as useActiveTradingAccount } from './useActiveTradingAccount';
 export { default as useActiveWalletAccount } from './useActiveWalletAccount';
+export { default as useActiveWalletBalance } from './useActiveWalletBalance';
 export { default as useAllAvailableAccounts } from './useAllAvailableAccounts';
 export { default as useAuthentication } from './useAuthentication';
 export { default as useAuthorize } from './useAuthorize';
@@ -37,7 +38,6 @@ export { default as usePOI } from './usePOI';
 export { default as useResidenceList } from './useResidenceList';
 export { default as useSettings } from './useSettings';
 export { default as useSortedMT5Accounts } from './useSortedMT5Accounts';
-export { default as useSubscribedBalance } from './useSubscribedBalance';
 export { default as useTradingAccountsList } from './useTradingAccountsList';
 export { default as useTradingPlatformInvestorPasswordChange } from './useTradingPlatformInvestorPasswordChange';
 export { default as useTradingPlatformInvestorPasswordReset } from './useTradingPlatformInvestorPasswordReset';
